requirements.txt
psycopg2-binary
pgvector
requests
numpy
tiktoken
